requirements.txt
# Requirements for Dune Analytics Strategic Intelligence + Alchemy Real-Time Monitoring

# Core analytics
dune-client>=1.0.0
psycopg2-binary>=2.9.0
requests>=2.28.0
python-dotenv>=1.0.0

# Real-time monitoring
web3>=6.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Optional for backtesting
matplotlib>=3.7.0
